+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 - use new `/flags` endpoint instead of `/decide` ([#345](https://github.com/PostHog/posthog-ios/pull/345))
 
+## 3.25.0 - 2025-04-30
+
+- feat: add support for conditional survey questions ([#343](https://github.com/PostHog/posthog-ios/pull/343))
+
+## 3.24.3 - 2025-04-29
+
+- fix: feature flags not loading on sdk init ([#346](https://github.com/PostHog/posthog-ios/pull/346))
+
 ## 3.24.2 - 2025-04-24
 
 - fix: optional link in survey question type ([#341](https://github.com/PostHog/posthog-ios/pull/341))